Moves the task/s out of pod to provider. Adds GCP
