+++ kubesae/providers/aws.py
@@ -5,6 +5,7 @@ Provides helpful EKS and ECR utilities.
 
 import invoke
 from colorama import Style
+from kubesae.pod import fetch_namespace_var
 
 
 @invoke.task()
@@ -45,7 +46,72 @@ def configure_eks_kubeconfig(c, cluster=None, region=None):
         region = c.config.aws.get("region", "us-east-1")
     c.run(f"aws eks update-kubeconfig --name {cluster} --region {region}")
 
+@invoke.task(name="sync_media")
+def sync_media_tree(
+    c,
+    target_env="staging",
+    media_bucket="MEDIA_STORAGE_BUCKET_NAME",
+    acl="public-read",
+    local_target="./media",
+    dry_run=False,
+    delete=False,
+    local=False,
+):
+    """Sync an S3 media tree for a given environment/namespace to another. 
+
+    Args:
+        target_env   (string, required): A deployment host defined in ansible host_vars (e.g. "production", "staging", "dev"). DEFAULT: staging
+        media_bucket (string, required): The variable name for media defined in settings and host_vars. DEFAULT: MEDIA_STORAGE_BUCKET_NAME
+        acl          (string, required): Sets the access policy on each object. DEFAULT: public-read
+                                         Possible values: [
+                                            private, public-read, public-read-write, authenticated-read, 
+                                            aws-exec-read, bucket-owner-read,bucket-owner-full-control,
+                                            log-delivery-write
+                                        ]
+        local_target (string, optional): Sets a default target for local syncs
+        dry_run      (boolean, optional): Outputs the result to stdout without applying the action
+        delete       (boolean, optional): If set, deletes files on the target that do not exist on the source.
+        local        (boolean, optional): If set, syncs media files to the location defined by the "local_target" parameter.
+
+    Usage:
+        inv production aws.sync-media --dry-run: 
+            Will simulate a sync from production to staging using the s3 bucket defined in MEDIA_STORAGE_BUCKET with no acl applied.
+        
+        inv production aws.sync-media --dry-run --delete
+            Will display the files that will be deleted from the staging s3 bucket defined in MEDIA_STORAGE_BUCKET.
+        
+        inv production aws.sync-media --media-bucket="MEDIA" --acl public-read --delete
+            Will sync files from the s3 bucket defined in the environment variable "MEDIA" to a staging bucket with the acl of each object set to 'public-read', and
+            will delete objects on the staging bucket that do not exist on the production bucket.
+        
+        inv production aws.sync-media --local --delete
+            Will sync files from the production s3 bucket to "<PROJECT_ROOT>/media"
+    """
+
+    cc = invoke.context.Context()
+    cc.config.env = target_env
+    cc.config.namespace = f"{c.config.app}-{target_env}"
+    cc.config.container_name = c.config.container_name
+
+    source_media_name = fetch_namespace_var(
+        c, fetch_var=f"{media_bucket}"
+    ).stdout.strip()
+    target_media_name = fetch_namespace_var(
+        cc, fetch_var=f"{media_bucket}"
+    ).stdout.strip()
+
+    dr = ""
+    dl = ""
+    if dry_run:
+        dr = "--dryrun"
+    if delete:
+        dl = "--delete"
+    if local:
+        c.run(f"aws s3 sync --acl {acl} s3://{source_media_name} {local_target} {dr} {dl}")
+    c.run(f"aws s3 sync --acl {acl} s3://{source_media_name} s3://{target_media_name} {dr} {dl}")
+
 
 aws = invoke.Collection("aws")
 aws.add_task(aws_docker_login, "docker-login")
 aws.add_task(configure_eks_kubeconfig, "configure-eks-kubeconfig")
+aws.add_task(sync_media_tree, "sync_media")
