+++ kubesae/providers/gcp.py
@@ -5,6 +5,7 @@ Provides helpful utilities for working with kubernetes and the Google Container 
 
 import invoke
 from colorama import Style
+from kubesae.pod import fetch_namespace_var
 
 
 @invoke.task()
@@ -46,7 +47,54 @@ def configure_gcp_kubeconfig(c, cluster=None, region=None):
     c.run(f"gcloud config set project {c.config.app}")
     c.run(f"gcloud container clusters get-credentials --region={region} {cluster}")
 
+@invoke.task(name="sync_media")
+def sync_media_tree(
+    c,
+    target_env="staging",
+    media_bucket="MEDIA_STORAGE_BUCKET_NAME",
+    local_target=None,
+    dry_run=False,
+    delete=False,
+):
+    """Sync a gcloud media tree for a given environment/namespace to another. 
+
+    Args:
+        target_env   (string, required): A deployment host defined in ansible host_vars (e.g. "production", "staging", "dev"). DEFAULT: staging
+        media_bucket (string, required): The variable name for media defined in settings and host_vars. DEFAULT: MEDIA_STORAGE_BUCKET_NAME
+        local_target (string, optional): Sets a target directory for local syncs
+        dry_run      (boolean, optional): Outputs the result to stdout without applying the action
+        delete       (boolean, optional): If set, deletes files on the target that do not exist on the source.
+
+    Usage:
+
+    """
+
+    cc = invoke.context.Context()
+    cc.config.env = target_env
+    cc.config.namespace = f"{c.config.app}-{target_env}"
+    cc.config.container_name = c.config.container_name
+
+    source_media_name = fetch_namespace_var(
+        c, fetch_var=f"{media_bucket}"
+    ).stdout.strip()
+    target_media_name = fetch_namespace_var(
+        cc, fetch_var=f"{media_bucket}"
+    ).stdout.strip()
+
+    dr = ""
+    dl = ""
+    if dry_run:
+        dr = "-n"
+    if delete:
+        dl = "-d"
+
+    target_media_name = f"gs://{target_media_name}"
+    if local_target:
+        target_media_name = local_target
+    c.run(f"gsutil rsync -r {dr} {dl} gs://{source_media_name} {target_media_name}")
+
 
 gcp = invoke.Collection("gcp")
 gcp.add_task(gcp_docker_login, "docker-login")
 gcp.add_task(configure_gcp_kubeconfig, "configure-gcp-kubeconfig")
+gcp.add_task(sync_media_tree)
